# MicroServ_Usuariosv3/src/main/java/org/openapitools/api/UserApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.9.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.User;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2024-10-15T15:55:01.292135700+02:00[Europe/Madrid]", comments = "Generator version: 7.9.0")
@Validated
@Tag(name = "user", description = "Operaciones sobre los usuarios")
public interface UserApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /user : Crea un usuario
     * Solo se podra hacer si el usuario ha iniciado sesion.
     *
     * @param user Crea un objeto User (optional)
     * @return operacion exitosa (status code 200)
     */
    @Operation(
        operationId = "createUser",
        summary = "Crea un usuario",
        description = "Solo se podra hacer si el usuario ha iniciado sesion.",
        tags = { "user" },
        responses = {
            @ApiResponse(responseCode = "default", description = "operacion exitosa", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = User.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/user",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<User> createUser(
        @Parameter(name = "User", description = "Crea un objeto User") @Valid @RequestBody(required = false) User user
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"lastName\" : \"James\", \"country\" : \"Spain\", \"role\" : \"User\", \"birthdate\" : \"1990-01-01\", \"bio\" : \"Esta es la descripcion de user1\", \"uploadedVideos\" : [ 101, 101 ], \"watchedVideos\" : [ 202, 202 ], \"firstName\" : \"John\", \"profilePicture\" : \"https://example.com/images/user1.jpg\", \"password\" : \"12345\", \"followers\" : [ 15, 15 ], \"following\" : [ 20, 20 ], \"id\" : 10, \"email\" : \"john@email.com\", \"username\" : \"user1\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /user/{username} : Borrar usuario
     * Solo se puede hacer si el usuario ha iniciado sesion.
     *
     * @param username El nombre del usuario a borrar (required)
     * @return operacion exitosa (status code 200)
     *         or Invalid username supplied (status code 400)
     *         or User not found (status code 404)
     */
    @Operation(
        operationId = "deleteUser",
        summary = "Borrar usuario",
        description = "Solo se puede hacer si el usuario ha iniciado sesion.",
        tags = { "user" },
        responses = {
            @ApiResponse(responseCode = "200", description = "operacion exitosa"),
            @ApiResponse(responseCode = "400", description = "Invalid username supplied"),
            @ApiResponse(responseCode = "404", description = "User not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/user/{username}"
    )
    
    default ResponseEntity<Void> deleteUser(
        @Parameter(name = "username", description = "El nombre del usuario a borrar", required = true, in = ParameterIn.PATH) @PathVariable("username") String username
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /user/{username}/followers : Se obtiene una lista de seguidores del usuario.
     * 
     *
     * @param username El nombre por el que se requieren sus seguidores  (required)
     * @return operacion exitosa (status code 200)
     */
    @Operation(
        operationId = "getFollowers",
        summary = "Se obtiene una lista de seguidores del usuario.",
        description = "",
        tags = { "user" },
        responses = {
            @ApiResponse(responseCode = "default", description = "operacion exitosa")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/user/{username}/followers"
    )
    
    default ResponseEntity<Void> getFollowers(
        @Parameter(name = "username", description = "El nombre por el que se requieren sus seguidores ", required = true, in = ParameterIn.PATH) @PathVariable("username") String username
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /user/{username}/following : Se obtiene una lista de usuarios seguidos por el usuario.
     * 
     *
     * @param username El nombre por el que se los usuarios que sigue  (required)
     * @return operacion exitosa (status code 200)
     */
    @Operation(
        operationId = "getFollowing",
        summary = "Se obtiene una lista de usuarios seguidos por el usuario.",
        description = "",
        tags = { "user" },
        responses = {
            @ApiResponse(responseCode = "default", description = "operacion exitosa")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/user/{username}/following"
    )
    
    default ResponseEntity<Void> getFollowing(
        @Parameter(name = "username", description = "El nombre por el que se los usuarios que sigue ", required = true, in = ParameterIn.PATH) @PathVariable("username") String username
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /user/{username} : Hacer un get del usuario por el nombre
     * 
     *
     * @param username El nombre por el que se busca. Usar user1 para testing.  (required)
     * @return operacion exitosa (status code 200)
     *         or Invalid username supplied (status code 400)
     *         or User not found (status code 404)
     */
    @Operation(
        operationId = "getUserByName",
        summary = "Hacer un get del usuario por el nombre",
        description = "",
        tags = { "user" },
        responses = {
            @ApiResponse(responseCode = "200", description = "operacion exitosa", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = User.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid username supplied"),
            @ApiResponse(responseCode = "404", description = "User not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/user/{username}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<User> getUserByName(
        @Parameter(name = "username", description = "El nombre por el que se busca. Usar user1 para testing. ", required = true, in = ParameterIn.PATH) @PathVariable("username") String username
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"lastName\" : \"James\", \"country\" : \"Spain\", \"role\" : \"User\", \"birthdate\" : \"1990-01-01\", \"bio\" : \"Esta es la descripcion de user1\", \"uploadedVideos\" : [ 101, 101 ], \"watchedVideos\" : [ 202, 202 ], \"firstName\" : \"John\", \"profilePicture\" : \"https://example.com/images/user1.jpg\", \"password\" : \"12345\", \"followers\" : [ 15, 15 ], \"following\" : [ 20, 20 ], \"id\" : 10, \"email\" : \"john@email.com\", \"username\" : \"user1\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /user/login : Se inicia sesion en el sistema
     * 
     *
     * @param username El nombre de usuario para iniciar sesion (optional)
     * @param password Contrasena necesaria para el login (optional)
     * @return operacion exitosa (status code 200)
     *         or Nombre de usuario o contrasena no valido (status code 400)
     */
    @Operation(
        operationId = "loginUser",
        summary = "Se inicia sesion en el sistema",
        description = "",
        tags = { "user" },
        responses = {
            @ApiResponse(responseCode = "200", description = "operacion exitosa", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = String.class))
            }),
            @ApiResponse(responseCode = "400", description = "Nombre de usuario o contrasena no valido")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/user/login",
        produces = { "application/json" }
    )
    
    default ResponseEntity<String> loginUser(
        @Parameter(name = "username", description = "El nombre de usuario para iniciar sesion", in = ParameterIn.QUERY) @Valid @RequestParam(value = "username", required = false) String username,
        @Parameter(name = "password", description = "Contrasena necesaria para el login", in = ParameterIn.QUERY) @Valid @RequestParam(value = "password", required = false) String password
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /user/logout : Se cierra sesion en el usuario actual
     * 
     *
     * @return operacion exitosa (status code 200)
     */
    @Operation(
        operationId = "logoutUser",
        summary = "Se cierra sesion en el usuario actual",
        description = "",
        tags = { "user" },
        responses = {
            @ApiResponse(responseCode = "default", description = "operacion exitosa")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/user/logout"
    )
    
    default ResponseEntity<Void> logoutUser(
        
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /user/{username} : Actualizar usuario
     * Solo puede utilizarse si el usuario ha iniciado sesion.
     *
     * @param username el nombre del usuario a modificar (required)
     * @param user Modifica el usuario en la aplicacion (optional)
     * @return operacion exitosa (status code 200)
     *         or Invalid username supplied (status code 400)
     *         or User not found (status code 404)
     */
    @Operation(
        operationId = "updateUser",
        summary = "Actualizar usuario",
        description = "Solo puede utilizarse si el usuario ha iniciado sesion.",
        tags = { "user" },
        responses = {
            @ApiResponse(responseCode = "200", description = "operacion exitosa"),
            @ApiResponse(responseCode = "400", description = "Invalid username supplied"),
            @ApiResponse(responseCode = "404", description = "User not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/user/{username}",
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Void> updateUser(
        @Parameter(name = "username", description = "el nombre del usuario a modificar", required = true, in = ParameterIn.PATH) @PathVariable("username") String username,
        @Parameter(name = "User", description = "Modifica el usuario en la aplicacion") @Valid @RequestBody(required = false) User user
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}